[util] factorize utils method stdVecToMatrice
